requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
PyJWT==2.9.0
djangorestframework_simplejwt==5.5.0
